src/backend/requirements.txt
flask
python-dotenv
psycopg2-binary
redis
gunicorn
